move cell languages code into notebooks
