+++ src/notebooks/languages/cellLanguageService.ts
@@ -6,16 +6,16 @@
 import type * as nbformat from '@jupyterlab/nbformat';
 import { inject, injectable, named } from 'inversify';
 import { Memento, NotebookCellKind, NotebookDocument } from 'vscode';
-import { IExtensionSingleActivationService } from '../platform/activation/types';
-import { IJupyterKernelSpec } from '../platform/api/extension';
-import { IPythonExtensionChecker } from '../platform/api/types';
-import { IVSCodeNotebook } from '../platform/common/application/types';
-import { PYTHON_LANGUAGE } from '../platform/common/constants';
-import { traceWarning } from '../platform/logging';
-import { IDisposableRegistry, IMemento, GLOBAL_MEMENTO } from '../platform/common/types';
-import { swallowExceptions } from '../platform/common/utils/decorators';
-import { getLanguageInNotebookMetadata } from '../kernels/helpers';
-import { getNotebookMetadata, isJupyterNotebook, translateKernelLanguageToMonaco } from '../platform/common/utils';
+import { IExtensionSingleActivationService } from '../../platform/activation/types';
+import { IJupyterKernelSpec } from '../../platform/api/extension';
+import { IPythonExtensionChecker } from '../../platform/api/types';
+import { IVSCodeNotebook } from '../../platform/common/application/types';
+import { PYTHON_LANGUAGE } from '../../platform/common/constants';
+import { traceWarning } from '../../platform/logging';
+import { IDisposableRegistry, IMemento, GLOBAL_MEMENTO } from '../../platform/common/types';
+import { swallowExceptions } from '../../platform/common/utils/decorators';
+import { getLanguageInNotebookMetadata } from '../../kernels/helpers';
+import { getNotebookMetadata, isJupyterNotebook, translateKernelLanguageToMonaco } from '../../platform/common/utils';
 
 export const LastSavedNotebookCellLanguage = 'DATASCIENCE.LAST_SAVED_CELL_LANGUAGE';
 /**
