+++ src/notebooks/languages/emptyNotebookCellLanguageService.ts
@@ -3,16 +3,16 @@
 
 import { inject, injectable } from 'inversify';
 import { languages, NotebookCellKind, NotebookDocument } from 'vscode';
-import { IExtensionSingleActivationService } from '../platform/activation/types';
-import { IVSCodeNotebook } from '../platform/common/application/types';
-import { PYTHON_LANGUAGE } from '../platform/common/constants';
-import { traceError } from '../platform/logging';
-import { IDisposableRegistry } from '../platform/common/types';
-import { noop } from '../platform/common/utils/misc';
-import { chainWithPendingUpdates } from '../kernels/execution/notebookUpdater';
-import { INotebookControllerManager } from '../notebooks/types';
-import { isJupyterNotebook, translateKernelLanguageToMonaco } from '../platform/common/utils';
-import { IVSCodeNotebookController } from '../notebooks/controllers/types';
+import { IExtensionSingleActivationService } from '../../platform/activation/types';
+import { IVSCodeNotebook } from '../../platform/common/application/types';
+import { PYTHON_LANGUAGE } from '../../platform/common/constants';
+import { traceError } from '../../platform/logging';
+import { IDisposableRegistry } from '../../platform/common/types';
+import { noop } from '../../platform/common/utils/misc';
+import { chainWithPendingUpdates } from '../../kernels/execution/notebookUpdater';
+import { INotebookControllerManager } from '../types';
+import { isJupyterNotebook, translateKernelLanguageToMonaco } from '../../platform/common/utils';
+import { IVSCodeNotebookController } from '../controllers/types';
 /**
  * If user creates a blank notebook, then they'll mostl likely end up with a blank cell with language, lets assume `Python`.
  * Now if the user changes the kernel to say `Julia`. After this, they need to also change the language of the cell.
