+++ src/notebooks/notebookCommandListener.ts
@@ -18,7 +18,7 @@ import { IVSCodeNotebook, ICommandManager, IApplicationShell } from '../platform
 import { IConfigurationService, IDataScienceCommandListener, IDisposableRegistry } from '../platform/common/types';
 import { Commands } from '../platform/common/constants';
 import { noop } from '../platform/common/utils/misc';
-import { NotebookCellLanguageService } from '../intellisense/cellLanguageService';
+import { NotebookCellLanguageService } from './languages/cellLanguageService';
 import { DisplayOptions } from '../kernels/displayOptions';
 import { IKernel, IKernelProvider } from '../kernels/types';
 import { getDisplayPath } from '../platform/common/platform/fs-paths';
